requirements.web.txt
Django==4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-cleanup==9.0.0
django-tinymce==4.1.0
pymysql==1.1.0
channels==4.3.1
daphne==4.2.1
channels_redis==4.3.0
redis==7.0.1

mysqlclient==2.2.7
python-dotenv==1.1.1

openai==1.107.1         # nếu huynh dùng chatbox
requests==2.32.5
PyJWT==2.10.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
